refactor: move Search component to Catalog.tsx
